# Conflicts:
#	src/utils.py

+++ src/data.py
@@ -143,14 +143,14 @@ class ClassifierDataset(Dataset):
         """
         if self.mode == "train":
             if self.margin_ratio != 0:
-                margin_len_begin = (self.seq_length * self.data_sample_rate) * self.margin_ratio
-                margin_len_end = (self.seq_length * self.data_sample_rate) * (1 - self.margin_ratio)
+                margin_len_begin = int((self.seq_length * self.data_sample_rate) * self.margin_ratio)
+                margin_len_end = int((self.seq_length * self.data_sample_rate) * (1 - self.margin_ratio))
                 start_time = random.randint(begin_time - margin_len_begin, end_time - margin_len_end)
             else:
-                start_time = random.randint(begin_time, end_time - self.seq_length * self.data_sample_rate)
+                start_time = random.randint(begin_time, end_time - int(self.seq_length * self.data_sample_rate))
         else:
             start_time = begin_time
-        data, _ = sf.read(path_to_file, start=start_time, stop=start_time + self.seq_length * self.data_sample_rate)
+        data, _ = sf.read(path_to_file, start=start_time, stop=start_time + int(self.seq_length * self.data_sample_rate))
         audio = torch.tensor(data, dtype=torch.float).unsqueeze(0)
         return audio
 
@@ -277,7 +277,7 @@ class InferenceDataset(Dataset):
         audio - pytorch tensor (1-D array)
         """
         data, _ = sf.read(self.file_path, start=begin_time,
-                          stop=begin_time + self.seq_length * self.data_sample_rate)
+                          stop=begin_time + int(self.seq_length * self.data_sample_rate))
         audio = torch.tensor(data, dtype=torch.float).unsqueeze(0)
         return audio
